+++ packages/react-native/src/components/index.ts
@@ -12,7 +12,7 @@ export type {
   ControlsContainerProps,
   FullscreenButtonProps,
   PlayButtonProps,
-  PlayerObjectFit,
+  ObjectFit,
   PlayerProps,
   ProgressProps,
   TitleProps,